Use the thread safe form of TH*::Fit
The version of TH*::Fit which uses a string to designate the function
to use is not thread safe. The string is used to lookup a global
instance of the TF1 and that instance can be used simultaneously
but multiple Fits to store their results. However, the version of
TH*::Fit which takes an explicit pointer to a TF1 is safe.